fleshing out the announcement, documents, and quizzes controllers and views

+++ app/helpers/announcements_helper.rb
@@ -1,10 +1,14 @@
 module AnnouncementsHelper
 
-  def instructor_links(announcement)
+  def announcement_title(announcement)
     returning(Array.new) do |html|
-      html << link_to(image_tag( 'comment_edit.png', :alt => 'Edit Announcement' ), edit_course_announcement_path(@course, announcement))
-      html << link_to(image_tag( 'comment_delete.png', :alt => 'Delete Announcement' ), course_announcement_path(@course, announcement),
-                       :method => :delete, :confirm => 'Are you sure?')
+      html << announcement.title
+
+      if current_user.instructor?
+        html << link_to(image_tag( 'comment_edit.png', :alt => 'Edit Announcement' ), edit_course_announcement_path(@course, announcement))
+        html << link_to(image_tag( 'comment_delete.png', :alt => 'Delete Announcement' ), course_announcement_path(@course, announcement),
+                        :method => :delete, :confirm => 'Are you sure?')
+      end
     end.join("\n")
   end
